PhabricatorPhabricatorAuthProvider: Remove unnecessary Remarkup
Summary:
This is a small fix in a specific user message example. See other messages in the very same file which also use `<tt>`.

Here the problematic phrase:

{F292970}

See the example in the phrase "Use lowercase letters, digits, and periods. For example: ...".

Test Plan: After this change, the specific message "The URI where the OAuth server is installed. For example" from the page `/auth/config/edit/?provider=PhabricatorPhabricatorAuthProvider` does not have anymore backticks shown to the end user.

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: speck, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D25232