Mobile: hide unuseful "Persistent Chat" checkbox
Summary:
On mobile devices like tablets or toasters the "Persistent Chat"
floating widget is already hidden.

So, the related checkbox available from the top navigation bar
is just confusing on tablet and mobile devices / toasters, since
that nice checkbox does nothing there.

On mobile and tablet, this is the graphical change:

| Before   | After     |
|----------|-----------|
|{F281239} | {F281235} |

This change do not change anything for desktop devices.
So, on desktop, that checkbox is obviously still visible.

Closes T15240

Test Plan:
- test on tablet and below: now the checkbox should be not visible
- test on desktop: the checkbox should still be visible

Reviewers: O1 Blessed Committers, Cigaryno, bfs, speck

Reviewed By: O1 Blessed Committers, Cigaryno, bfs, speck

Subscribers: avivey, bfs, dcog, chris, speck, tobiaswiese, Matthew, Cigaryno

Maniphest Tasks: T15240

Differential Revision: https://we.phorge.it/D25120